+++ src/services/users/index.ts
@@ -5,6 +5,8 @@ import httpClient from '../http-client';
 import globalRoles from './globalRoles';
 import groupRoles from './groupRoles';
 import health from './health';
+import { userSettingsService } from './settings';
+import { UsersSettingsService } from './settings/types';
 import { UsersGlobalRolesService, UsersGroupRolesService } from './types';
 import users from './users';
 
@@ -15,6 +17,7 @@ export const usersService = (
   ReturnType<typeof health> & {
     globalRoles: UsersGlobalRolesService;
     groupRoles: UsersGroupRolesService;
+    settings: UsersSettingsService;
   } => {
   const userClient = httpClient({
     basePath: USER_BASE,
@@ -25,11 +28,13 @@ export const usersService = (
   const usersMethods = users(userClient, httpWithAuth, http);
   const groupRolesMethods = groupRoles(userClient, httpWithAuth);
   const globalRolesMethods = globalRoles(userClient, httpWithAuth);
+  const settingsMethods = userSettingsService(userClient, httpWithAuth);
 
   return {
     ...healthMethods,
     ...usersMethods,
     groupRoles: groupRolesMethods,
     globalRoles: globalRolesMethods,
+    settings: settingsMethods,
   };
 };